Fix report deferral logic, empty pool checks, and header includes in ReportScheduler
- Ensure DeferReports only updates deferral timestamp if new deferral end is greater than GetMinTimestamp()
- Add empty pool check in SynchronizedReportSchedulerImpl::RescheduleAllReports()
- Fix unused parameter warning in TestCommandInteraction.cpp
- Explicitly include <lib/support/Span.h> and update mNumTargetedEndpoints type to uint16_t in CommandHandlerImpl.h
- Add unit test verifying shorter deferral delay does not shrink existing deferral window

+++ src/app/CommandHandlerImpl.h
@@ -30,6 +30,7 @@
 #include <lib/core/TLVDebug.h>
 #include <lib/support/BitFlags.h>
 #include <lib/support/Scoped.h>
+#include <lib/support/Span.h>
 #include <messaging/ExchangeHolder.h>
 #include <messaging/Flags.h>
 #include <protocols/Protocols.h>
@@ -513,7 +514,7 @@ private:
     bool mGoneAsync = false;
 
     static constexpr size_t kMaxTargetedEndpoints = CHIP_CONFIG_MAX_PATHS_PER_INVOKE;
-    uint8_t mNumTargetedEndpoints                 = 0;
+    uint16_t mNumTargetedEndpoints                = 0;
     EndpointId mTargetedEndpoints[kMaxTargetedEndpoints];
     void RecordTargetedEndpoint(EndpointId endpointId);
 };
